+++ contrib/alexbeloi/examples/vpgis_cartpole.py
@@ -6,8 +6,8 @@ from contrib.alexbeloi.is_sampler import ISSampler
 from garage.baselines import LinearFeatureBaseline
 from garage.envs import normalize
 from garage.envs.box2d import CartpoleEnv
-from garage.policies import GaussianMLPPolicy
 from garage.theano.algos import VPG
+from garage.theano.policies import GaussianMLPPolicy
 
 env = normalize(CartpoleEnv())
