+++ contrib/ros/envs/sawyer/launchers/trpo_real_sawyer_push.py
@@ -5,8 +5,8 @@ from contrib.ros.envs.sawyer.push_env import PushEnv
 from garage.baselines.linear_feature_baseline import LinearFeatureBaseline
 from garage.envs.util import spec
 from garage.misc.instrument import run_experiment
-from garage.policies.gaussian_mlp_policy import GaussianMLPPolicy
 from garage.theano.algos import TRPO
+from garage.theano.policies import GaussianMLPPolicy
 
 INITIAL_ROBOT_JOINT_POS = {
     'right_j0': -0.140923828125,
