+++ contrib/ros/envs/sawyer/launchers/trpo_sim_sawyer_pnp.py
@@ -5,8 +5,8 @@ from contrib.ros.envs.sawyer.pick_and_place_env import PickAndPlaceEnv
 from garage.baselines.linear_feature_baseline import LinearFeatureBaseline
 from garage.envs.util import spec
 from garage.misc.instrument import run_experiment
-from garage.policies.gaussian_mlp_policy import GaussianMLPPolicy
 from garage.theano.algos import TRPO
+from garage.theano.policies import GaussianMLPPolicy
 
 INITIAL_ROBOT_JOINT_POS = {
     'right_j0': -0.140923828125,
