+++ contrib/ros/envs/sawyer/launchers/trpo_sim_sawyer_reacher.py
@@ -10,8 +10,8 @@ from contrib.ros.envs.sawyer import ReacherEnv
 from garage.baselines import LinearFeatureBaseline
 from garage.envs.util import spec
 from garage.misc.instrument import run_experiment
-from garage.policies import GaussianMLPPolicy
 from garage.theano.algos import TRPO
+from garage.theano.policies import GaussianMLPPolicy
 
 INITIAL_ROBOT_JOINT_POS = {
     'right_j0': -0.140923828125,
