+++ examples/theano/trpo_cartpole.py
@@ -1,9 +1,9 @@
 from garage.baselines import LinearFeatureBaseline
 from garage.envs import normalize
 from garage.envs.box2d import CartpoleEnv
-from garage.policies import GaussianMLPPolicy
 from garage.theano.algos import TRPO
 from garage.theano.envs import TheanoEnv
+from garage.theano.policies import GaussianMLPPolicy
 
 env = TheanoEnv(normalize(CartpoleEnv()))
