+++ examples/theano/trpo_gym_cartpole.py
@@ -3,9 +3,9 @@ import gym
 from garage.baselines import LinearFeatureBaseline
 from garage.envs import normalize
 from garage.misc.instrument import run_experiment
-from garage.policies import CategoricalMLPPolicy
 from garage.theano.algos import TRPO
 from garage.theano.envs import TheanoEnv
+from garage.theano.policies import CategoricalMLPPolicy
 
 
 def run_task(*_):
