+++ garage/policies/__init__.py
@@ -1,8 +1,2 @@
 from garage.policies.base import Policy
 from garage.policies.base import StochasticPolicy
-from garage.policies.categorical_gru_policy import CategoricalGRUPolicy
-from garage.policies.categorical_mlp_policy import CategoricalMLPPolicy
-from garage.policies.deterministic_mlp_policy import DeterministicMLPPolicy
-from garage.policies.gaussian_gru_policy import GaussianGRUPolicy
-from garage.policies.gaussian_mlp_policy import GaussianMLPPolicy
-from garage.policies.uniform_control_policy import UniformControlPolicy
