+++ garage/theano/policies/__init__.py
@@ -1,0 +1,7 @@
+from garage.theano.policies.categorical_gru_policy import CategoricalGRUPolicy
+from garage.theano.policies.categorical_mlp_policy import CategoricalMLPPolicy
+from garage.theano.policies.deterministic_mlp_policy import \
+        DeterministicMLPPolicy
+from garage.theano.policies.gaussian_gru_policy import GaussianGRUPolicy
+from garage.theano.policies.gaussian_mlp_policy import GaussianMLPPolicy
+from garage.theano.policies.uniform_control_policy import UniformControlPolicy
