+++ garage/theano/policies/categorical_conv_policy.py
