+++ garage/theano/policies/categorical_gru_policy.py
