+++ garage/theano/policies/categorical_mlp_policy.py
