+++ garage/theano/policies/gaussian_gru_policy.py
