+++ garage/theano/policies/uniform_control_policy.py
