+++ tests/envs/test_sawyer_envs.py
@@ -10,8 +10,8 @@ from garage.envs.mujoco.sawyer import BlockStackingEnv
 from garage.envs.mujoco.sawyer import PickAndPlaceEnv
 from garage.envs.mujoco.sawyer import ReacherEnv
 from garage.envs.util import spec
-from garage.policies import GaussianMLPPolicy
 from garage.theano.algos import TRPO
+from garage.theano.policies import GaussianMLPPolicy
 
 
 def run_bin_sorting(*_):
