+++ tests/test_adaptive_std.py
@@ -2,9 +2,9 @@ from nose2.tools import such
 
 from garage.baselines import ZeroBaseline
 from garage.envs.box2d import CartpoleEnv
-from garage.policies import GaussianMLPPolicy
 from garage.theano.algos import TRPO
 from garage.theano.envs import TheanoEnv
+from garage.theano.policies import GaussianMLPPolicy
 
 with such.A("adaptive_std") as it:
